requirements.txt
py-cord==2.7.0rc2
python-dotenv==1.0.1
PyNaCl==1.5.0
imageio-ffmpeg==0.4.9
aiohttp
httpx
motor==3.6.0
pymongo==4.9.2
dnspython==2.7.0
tzdata